create kernel directory
