+++ kernel/Makefile
@@ -2,7 +2,9 @@ include ../make.config
 
 SRC_DIR		:= src
 OBJ_DIR		:= obj
-INCLD_DIR	:= include
+INCLUDES	:= include ../$(LIBK_DIR)/include
+INCLD_FLAG	:= $(addprefix -I , $(INCLUDES))
+LIB			:= ../$(LIBK_DIR)/$(LIBK)
 
 SRC_C 		:= $(shell find $(SRC_DIR) -type f -name '*.c')
 SRC_S		:= $(shell find $(SRC_DIR) -type f -name '*.s')
@@ -10,18 +12,19 @@ OBJ			:= $(patsubst $(SRC_DIR)/%.c, $(OBJ_DIR)/%.c.o, $(SRC_C))
 OBJ			+= $(patsubst $(SRC_DIR)/%.s, $(OBJ_DIR)/%.s.o, $(SRC_S))
 DEP			:= $(OBJ:.o=.d)
 
+.PHONY: all clean re
+
 all: $(NAME_BIN)
 
 $(NAME_BIN): kos.ld $(OBJ)
-	$(CC) -T kos.ld -o $(NAME_BIN) $(CFLAGS) $(OBJ) -lgcc
+	$(CC) -T kos.ld -o $(NAME_BIN) $(CFLAGS) $(INCLD_FLAG) $(OBJ) $(LIB) -lgcc
 
-$(OBJ_DIR):
-	mkdir -p $@
+$(OBJ_DIR)/%.c.o: $(SRC_DIR)/%.c
+	@mkdir -p $(@D)
+	$(CC) -MD $(CFLAGS) $(INCLD_FLAG) $(LIB) -c $^ -o $@
 
-$(OBJ_DIR)/%.c.o: $(SRC_DIR)/%.c | $(OBJ_DIR)
-	$(CC) -MD $(CFLAGS) -I$(INCLD_DIR) -c $^ -o $@
-
-$(OBJ_DIR)/%.s.o: $(SRC_DIR)/%.s | $(OBJ_DIR)
+$(OBJ_DIR)/%.s.o: $(SRC_DIR)/%.s
+	@mkdir -p $(@D)
 	$(AS) $(AFLAGS) $^ -o $@
 
 clean:
