+++ make.config
@@ -4,10 +4,11 @@ NAME				= kos
 NAME_BIN			= $(NAME).bin
 NAME_ISO			= $(NAME).iso
 
-KERNEL				= kernel
-LIBC				= libc
 ROOT_DIR			= .
 ISO_DIR				= isodir
+KERNEL_DIR			= kernel
+LIBK_DIR			= libk
+LIBK				= libk.a
 GRUB_CFG			= grub.cfg
 
 CC					= $(HOST)-gcc
